cred: separate public interface from low-level details
