+++ tests/network/cred.c
@@ -23,28 +23,24 @@ void test_network_cred__stock_userpass_validates_that_method_is_allowed(void)
 
 	cl_git_fail(git_cred_userpass(&cred, NULL, NULL, 0, &payload));
 	cl_git_pass(git_cred_userpass(&cred, NULL, NULL, GIT_CREDTYPE_USERPASS_PLAINTEXT, &payload));
-	cred->free(cred);
+	git_cred_free(cred);
 }
 
 void test_network_cred__stock_userpass_properly_handles_username_in_url(void)
 {
 	git_cred *cred;
-	git_cred_userpass_plaintext *plain;
 	git_cred_userpass_payload payload = {"alice", "password"};
 
 	cl_git_pass(git_cred_userpass(&cred, NULL, NULL, GIT_CREDTYPE_USERPASS_PLAINTEXT, &payload));
-	plain = (git_cred_userpass_plaintext*)cred;
-	cl_assert_equal_s(plain->username, "alice");
-	cred->free(cred);
+	cl_assert_equal_s("alice", git_cred_get_username(cred));
+	git_cred_free(cred);
 
 	cl_git_pass(git_cred_userpass(&cred, NULL, "bob", GIT_CREDTYPE_USERPASS_PLAINTEXT, &payload));
-	plain = (git_cred_userpass_plaintext*)cred;
-	cl_assert_equal_s(plain->username, "alice");
-	cred->free(cred);
+	cl_assert_equal_s("alice", git_cred_get_username(cred));
+	git_cred_free(cred);
 
 	payload.username = NULL;
 	cl_git_pass(git_cred_userpass(&cred, NULL, "bob", GIT_CREDTYPE_USERPASS_PLAINTEXT, &payload));
-	plain = (git_cred_userpass_plaintext*)cred;
-	cl_assert_equal_s(plain->username, "bob");
-	cred->free(cred);
+	cl_assert_equal_s("bob", git_cred_get_username(cred));
+	git_cred_free(cred);
 }
